+++ src/pages/givingform.tsx
@@ -2,12 +2,12 @@ import React from "react"
 import Layout from "../components/layout"
 import { Form } from "../components/form/form"
 import { ValueIn, ValueEqual } from "../components/form/form-config.types"
-import { FormSectionStart } from "../components/form/form-part"
-import BasicTextField from "../components/form/basic-text-field"
-import RadioButtonField from "../components/form/radio-button-field"
-import DropDownField from "../components/form/drop-down-field"
-import DateField from "../components/form/date-field"
-import CheckBoxField from "../components/form/checkbox-field"
+import { FormSectionStart } from "../components/form/fields/form-section-start"
+import BasicTextField from "../components/form/fields/basic-text-field"
+import RadioButtonField from "../components/form/fields/radio-button-field"
+import DropDownField from "../components/form/fields/drop-down-field"
+import DateField from "../components/form/fields/date-field"
+import CheckBoxField from "../components/form/fields/checkbox-field"
 import { ValidationOptions } from "react-hook-form"
 import {
     convertFormDateToGoogleFormUrl,
